% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_articles}
\alias{get_articles}
\title{Get Data}
\usage{
get_articles(q, since, until, language = NULL, results = 1000)
}
\arguments{
\item{q}{The terms to query.}

\item{since, until}{Date range to query.}

\item{language}{Language to limit results.}

\item{results}{Number of results to return.}
}
\description{
Get data from both data sources.
}
