% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{letterbox_keys}
\alias{letterbox_keys}
\title{Setup}
\usage{
letterbox_keys(currents_key = NULL, newsapi_key = NULL)
}
\arguments{
\item{currents_key, newsapi_key}{The relevant API keys, \url{newsapi.org} and \url{currentsapi.services}.}
}
\description{
Setup the necessary keys.
}
